feat: remove useless deps and simplify some code
