+++ kernel/Cargo.toml
@@ -23,7 +23,6 @@ libc-types = { workspace = true }
 xmas-elf = "0.9.0"
 async-recursion = "1.1.1"
 sync = { workspace = true }
-bitflags = "2.0.2"
 bit_field = "0.10.1"
 lose-net-stack = { git = "https://github.com/byte-os/lose-net-stack", rev = "bb99460", features = [
   "log",
